+++ apps/ui-dev/component/main/data.jsx
@@ -12,25 +12,25 @@ export const data = [
       "https://github.com/elf-framework/editor/tree/develop/packages/sapa",
   },
 
-  {
-    title: "@elf-framework/design-tokens",
-    npm: "@elf-framework/design-tokens",
-    description: "ELF Design Tokens",
-    docLink: url("pages/design-system/getting-started/index.html"),
-    codeLinkText: "Github",
-    codeLink:
-      "https://github.com/elf-framework/editor/tree/develop/packages/design-tokens",
-  },
+  // {
+  //   title: "@elf-framework/design-tokens",
+  //   npm: "@elf-framework/design-tokens",
+  //   description: "ELF Design Tokens",
+  //   docLink: url("pages/design-system/getting-started/index.html"),
+  //   codeLinkText: "Github",
+  //   codeLink:
+  //     "https://github.com/elf-framework/editor/tree/develop/packages/design-tokens",
+  // },
 
-  {
-    title: "@elf-framework/design-system",
-    npm: "@elf-framework/design-system",
-    description: "ELF Design System",
-    docLink: url("pages/design-system/getting-started/index.html"),
-    codeLinkText: "Github",
-    codeLink:
-      "https://github.com/elf-framework/editor/tree/develop/packages/design-system",
-  },
+  // {
+  //   title: "@elf-framework/design-system",
+  //   npm: "@elf-framework/design-system",
+  //   description: "ELF Design System",
+  //   docLink: url("pages/design-system/getting-started/index.html"),
+  //   codeLinkText: "Github",
+  //   codeLink:
+  //     "https://github.com/elf-framework/editor/tree/develop/packages/design-system",
+  // },
 
   {
     title: "@elf-framework/ui",
@@ -61,34 +61,34 @@ export const data = [
       "https://github.com/elf-framework/editor/tree/develop/packages/base-editor",
     type: "beta",
   },
-  {
-    title: "@elf-framework/color",
-    npm: "@elf-framework/color",
-    description: "ELF color library",
-    docLink: "https://www.elf-framework.com",
-    codeLinkText: "Github",
-    codeLink:
-      "https://github.com/elf-framework/editor/tree/develop/packages/color",
-  },
+  // {
+  //   title: "@elf-framework/color",
+  //   npm: "@elf-framework/color",
+  //   description: "ELF color library",
+  //   docLink: "https://www.elf-framework.com",
+  //   codeLinkText: "Github",
+  //   codeLink:
+  //     "https://github.com/elf-framework/editor/tree/develop/packages/color",
+  // },
 
-  {
-    title: "@elf-framework/sapa-router",
-    npm: "@elf-framework/sapa-router",
-    description: "Client router for sapa",
-    docLink: "https://www.elf-framework.com",
-    codeLinkText: "Github",
-    codeLink:
-      "https://github.com/elf-framework/editor/tree/develop/packages/sapa-router",
-    type: "beta",
-  },
+  // {
+  //   title: "@elf-framework/sapa-router",
+  //   npm: "@elf-framework/sapa-router",
+  //   description: "Client router for sapa",
+  //   docLink: "https://www.elf-framework.com",
+  //   codeLinkText: "Github",
+  //   codeLink:
+  //     "https://github.com/elf-framework/editor/tree/develop/packages/sapa-router",
+  //   type: "beta",
+  // },
 
-  {
-    title: "vite-plugin-sapa",
-    npm: "vite-plugin-sapa",
-    description: "Vite plugin to support HMR for sapa",
-    docLink: "https://www.elf-framework.com",
-    codeLinkText: "Github",
-    codeLink:
-      "https://github.com/elf-framework/editor/tree/develop/packages/vite-plugin-sapa",
-  },
+  // {
+  //   title: "vite-plugin-sapa",
+  //   npm: "vite-plugin-sapa",
+  //   description: "Vite plugin to support HMR for sapa",
+  //   docLink: "https://www.elf-framework.com",
+  //   codeLinkText: "Github",
+  //   codeLink:
+  //     "https://github.com/elf-framework/editor/tree/develop/packages/vite-plugin-sapa",
+  // },
 ];